create_game(): Allow to specify a starting airport.
Instead of picking an airport at random to start form, callers may optionally
pass in a starting airport (AirportMaster object).  If specified, the starting
airport will be copied over from that one.

This is to be used in our new geo-ip based feature where the starting airport
will be in the city where the game host is when he/she creates the game (if
available).