Make new game's starting airport closes to the host.
This is like the previous feature of making it the host's city except it
doesn't tie it to the city.  Instead it uses geoip to find the city closes to
the host and then starts the game in an airport in that city.  That way you
don't actually have to be in a city that has an airport to feel the love.