game: change the way we create flights.
Ok, the way things used to work is, for a given airport, we create flights for
all the destinations.  When every flight has departed, then we re-create
flights.  This makes players have to wait until all flights are departed
before they can choose a different flight.

The way things work on this commit is different.  Instead of waiting for every
flight for the airport to depart, we instead do it on a per-destination basis,
and once per hour.  Meaning, if the airport is RDU and there is a flight
leaving for DFW at 4:10, then as soon as that flight departs we create a new
RDU->DFW flight.  However the new flight will depart no earlier than 5:00 and
no later than 5:59.  So you still make the player wait, but the flight
creation is staggered.  It also reduces the number of flights.  There is a
maximum of 24 flights per destination per day.

I've experimented with this and it *seems* good, but I may back it out later
if I find problems with it.