models.Airport.create_flights: Bring back the cushion
In cs 9fc60ae2d61f we changed the way flights were created to randomized
hourly flights.  This seems to work alright but has one disadvantage that the
previous hour's flight could have departed at, say 9:59 and the next hour's
flight could depart at 10:00, which is only a minute later.  This change
(re)adds a 20 minute cushion, so that if the aforementioned 9:59 flight
occurred then the next flight will not depart before 10:19.

I may decide to adjust the 20 minutes later (or make it a setting).  To me 30
mins felt like a little too much.