tests: Rewrite a test.
The next_flight_to() mostly passes, but occasionally fails.  It appears to be
an off-by-one (or rather off-by-30-minutes) issue, but I don't see it, so I
basically rewrote the test and made it a bit simpler.  Hopefully this fixes
the occasional failure (unless the failure is in the actual code and not the
test).