use a single-thread/process gameserver
This change was influenced by Jaime Bueleta's blog entry[1].  And while I read
it and don't anticipate the issues he lays out will ever be a problem for
Airport (it's basically a play-at-home game and not intended to be massively
consumed), I do at least agree with things on principle and, sinc it was not
difficult to do, I made the change.  Besides, I was kinda looking for a reason
to do something with Airport.

We probably don't even need to spawn a process/thread at all and, if this
turns out to be stable enough, later I may just make `GameThread` a
non-threaded class and just `run()` it as that.

[1] https://wrongsideofmemphis.wordpress.com/2015/12/16/do-not-spawn-processes-on-users-requests/